Copy edit for clarity and grammar
Changed "On your site or blog in no time. As easy as 1-2-3" to "TimelineJS works on any site or blog. Make your own in four easy steps"
Changed "Create a spreadsheet" to "Create your spreadsheet"
Changed "It's as simple as dropping dates ..." to "Drop dates, text and links in to the appropriate columns."