Updated iframe regex for MediaType detection
The previous regex required the iframe to have an attribute after "src", so <iframe src="..."></iframe> didn't work, but things like <iframe src="..." frameborder="no"></iframe> did. Since only the src is used and the rest of the iframe is reconstructed, there's no need for extra attributes, and had me searching for a bit.