LESS: fix control positioning on RTL pages
Currently the next and previous controls will overlap when embedded on an RTL page. This commit makes the controls usable but doesn’t start the larger project of flipping the various left/right classes applied to .nav-next / .nav-previous elements or updating the JavaScript to use “left” or “right” based on whether the timeline is embedded in a dir=rtl or dir=ltr element.