Prevents crashes...
I'm owrking on creating simple data filters for my timeline JS implementation.  so far it's been pretty easy I'm just filtering the data on my page and calling the reload method.  However, if you use the reload function and are on a slide the current slide will get jacked up and cause a crash.  This prevents the crash.